+++ src/main/java/com/quasiris/qsf/pipeline/filter/elastic/spellcheck/SpellCheckTokenType.java
@@ -1,0 +1,6 @@
+package com.quasiris.qsf.pipeline.filter.elastic.spellcheck;
+
+public enum SpellCheckTokenType {
+
+    CORRECTED, CORRECT, UNKNOWN
+}